samples: nrf9160: gps: fix printf va_args
%lu is for long unsigned types; uint32_t on this platform is an
unsigned int. Add a cast to ensure the args are the right type.

+++ samples/nrf9160/gps/src/main.c
@@ -37,12 +37,12 @@ nrf_gnss_data_frame_t last_fix;
 
 void bsd_recoverable_error_handler(uint32_t error)
 {
-	printf("Err: %lu\n", error);
+	printf("Err: %lu\n", (unsigned long)error);
 }
 
 void bsd_irrecoverable_error_handler(uint32_t error)
 {
-	printf("Irrecoverable: %lu\n", error);
+	printf("Irrecoverable: %lu\n", (unsigned long)error);
 }
 
 static int enable_gps(void)
